+++ cmd/mysql.go
@@ -13,12 +13,12 @@ import (
 
 var (
 	MySQLConsoleCommand = cli.Command{
-		Name:     "mysql-console",
-		Category: "Databases",
+		Name:     mySQLConsole,
+		Category: categoryDatabases,
 		Usage:    "Run an interactive console with your MySQL addon",
 		Flags: []cli.Flag{&appFlag,
-			&cli.StringFlag{Name: "size", Aliases: []string{"s"}, Value: "", Usage: "Size of the container"},
-			&cli.StringFlag{Name: "env", Aliases: []string{"e"}, Value: "", Usage: "Environment variable name to use for the connection to the database"},
+			&cli.StringFlag{Name: flagSizeName, Aliases: []string{"s"}, Value: "", Usage: flagSizeUsage},
+			&cli.StringFlag{Name: flagEnvName, Aliases: []string{"e"}, Value: "", Usage: flagEnvUsage},
 		},
 		Description: CommandDescription{
 			Description: `Run an interactive console with your MySQL addon
@@ -26,17 +26,17 @@ var (
 The --size flag makes it easy to specify the size of the container executing
 the MySQL console. Each container size has different price and performance.
 You can read more about container sizes here:
-http://doc.scalingo.com/internals/container-sizes.html`,
+https://doc.scalingo.com/platform/internals/container-sizes`,
 			Examples: []string{
 				"scalingo --app my-app mysql-console",
 				"scalingo --app my-app mysql-console --size L",
 				"scalingo --app my-app mysql-console --env MY_MYSQL_URL",
 			},
-			SeeAlso: []string{"mongo-console", "pgsql-console"},
+			SeeAlso: []string{mongoConsole, pgSQLConsole},
 		}.Render(),
 		Action: func(ctx context.Context, c *cli.Command) error {
 			if c.Args().Len() != 0 {
-				_ = cli.ShowCommandHelp(ctx, c, "mysql-console")
+				_ = cli.ShowCommandHelp(ctx, c, mySQLConsole)
 				return nil
 			}
 
@@ -54,7 +54,7 @@ http://doc.scalingo.com/internals/container-sizes.html`,
 			return nil
 		},
 		ShellComplete: func(_ context.Context, c *cli.Command) {
-			_ = autocomplete.CmdFlagsAutoComplete(c, "mysql-console")
+			_ = autocomplete.CmdFlagsAutoComplete(c, mySQLConsole)
 		},
 	}
 )